Allow link-exporting into Trello/Notion
This fixes the link export issue in macOS Chrome (https://github.com/sunsama/sunsama/issues/7671), _but_ it effectively disables the Esc cancel as well.

I spent a couple of hours on this and couldn't come up with a better condition, and I think it's more important to allow canceling than to allow the link export with CMD in macOS Chrome, but still left this out in case @bone-saw wants to merge, test, or improve it.

Will need the `sunsama-app/package.json` change in version if it gets merged.